Organize commands into new folders
Makes name of mechanism (notemechanism) more consistent.
