+++ 2024-Competition-Code/src/main/java/frc/robot/RobotContainer.java
@@ -5,9 +5,9 @@ import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
-
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import edu.wpi.first.wpilibj2.command.button.POVButton;
+
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
@@ -16,6 +16,8 @@ import edu.wpi.first.wpilibj.DriverStation;
 import org.photonvision.PhotonCamera;
 
 import frc.robot.commands.*;
+import frc.robot.commands.notemechanism.*;
+import frc.robot.commands.swerve.*;
 import frc.robot.subsystems.*;
 import frc.robot.subsystems.swerve.*;
 import frc.robot.subsystems.photonvision.*;
