+++ 2024-Competition-Code/src/main/java/frc/robot/commands/notemechanism/IntakeBeamBreak.java
@@ -2,42 +2,43 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.notemechanism;
 
-import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.Shooter;
 
-public class IntakeManual extends Command {
-  /** Creates a new IntakeManual. */
+public class IntakeBeamBreak extends Command {
+  /** Creates a new IntakeCommand. */
   private double power;
   private Shooter shooter;
-  public IntakeManual(double power, Shooter shooter) {
-    // Use addRequirements() here to declare subsystem dependencies.
+  public IntakeBeamBreak(double power, Shooter shooter) {
     this.power = power;
     this.shooter = shooter;
     addRequirements(shooter);
+    // Use addRequirements() here to declare subsystem dependencies.
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {}
+  public void initialize() {
+  }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    shooter.setFeedSimple(power);
+    //shooter.setintakeSpeed(power);
+    shooter.setIntakePower(power);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    shooter.setFeedSimple(0);
+    shooter.setIntakePower(0);
   }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return false;
+    return shooter.getShooterBeam();
   }
 }
