+++ 2024-Competition-Code/src/main/java/frc/robot/commands/notemechanism/IntakeManual.java
@@ -2,23 +2,21 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.notemechanism;
 
+import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.Shooter;
 
-
-public class Shoot extends Command {
+public class IntakeManual extends Command {
+  /** Creates a new IntakeManual. */
   private double power;
-  private double lowerRPM;
   private Shooter shooter;
-  /** Creates a new Shoot. */
-  public Shoot(double power, double lowerRPM, Shooter shooter) {
+  public IntakeManual(double power, Shooter shooter) {
+    // Use addRequirements() here to declare subsystem dependencies.
     this.power = power;
     this.shooter = shooter;
-    this.lowerRPM = lowerRPM;
     addRequirements(shooter);
-    // Use addRequirements() here to declare subsystem dependencies.
   }
 
   // Called when the command is initially scheduled.
@@ -28,17 +26,13 @@ public class Shoot extends Command {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    shooter.setupperSpeed(power);
-    shooter.setlowerSpeed(lowerRPM);
-    // if (shooter.setpointReached(shooter.getUpperRPM(), shooter.upperRPM) && (shooter.setpointReached(shooter.getLowerRPM(), shooter.lowerRPM))){
-    //   shooter.setintakeSpeed(3000);
-    // }
+    shooter.setFeedSimple(power);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    shooter.setupperSpeed(0);
+    shooter.setFeedSimple(0);
   }
 
   // Returns true when the command should end.
