+++ 2024-Competition-Code/src/main/java/frc/robot/commands/notemechanism/Shoot.java
@@ -2,43 +2,48 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.notemechanism;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.Shooter;
 
-public class IntakeBeamBreak extends Command {
-  /** Creates a new IntakeCommand. */
+
+public class Shoot extends Command {
   private double power;
+  private double lowerRPM;
   private Shooter shooter;
-  public IntakeBeamBreak(double power, Shooter shooter) {
+  /** Creates a new Shoot. */
+  public Shoot(double power, double lowerRPM, Shooter shooter) {
     this.power = power;
     this.shooter = shooter;
+    this.lowerRPM = lowerRPM;
     addRequirements(shooter);
     // Use addRequirements() here to declare subsystem dependencies.
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {
-  }
+  public void initialize() {}
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    //shooter.setintakeSpeed(power);
-    shooter.setIntakePower(power);
+    shooter.setupperSpeed(power);
+    shooter.setlowerSpeed(lowerRPM);
+    // if (shooter.setpointReached(shooter.getUpperRPM(), shooter.upperRPM) && (shooter.setpointReached(shooter.getLowerRPM(), shooter.lowerRPM))){
+    //   shooter.setintakeSpeed(3000);
+    // }
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    shooter.setIntakePower(0);
+    shooter.setupperSpeed(0);
   }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return shooter.getShooterBeam();
+    return false;
   }
 }
