+++ 2024-Competition-Code/src/main/java/frc/robot/commands/swerve/SwerveTeleop.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.swerve;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.geometry.Translation2d;
